+++ test/state.c
@@ -774,7 +774,7 @@ test_update_key(struct xkb_context *ctx, struct xkb_keymap *keymap,
     struct xkb_machine *sm = xkb_machine_new(builder);
     assert(sm);
     xkb_machine_builder_destroy(builder);
-    struct xkb_events *events = xkb_events_new_batch(ctx, XKB_EVENTS_NO_FLAGS);
+    struct xkb_events *events = xkb_events_new(ctx, NULL, NULL);
     assert(events);
     const xkb_keysym_t *syms;
     xkb_keysym_t one_sym;